Fix namespaces after moving test files
